requirements.txt
backcall==0.2.0
colorama==0.4.3
decorator==4.4.2
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
numpy==1.19.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
Pygments==2.7.1
traitlets==5.0.4
wcwidth==0.2.5
